Implement flag commands to spawn and move
Changed the action store to contain just an array of objects. So the store is actually represented not by the current state, yet by the desired state, and when the store is empty, it means that the desired state is reached.  
ActionInitializer is responsible for achieving the desired state. It proceeds actions from the action store to meet the desired state.

FlagListener listens for new flags placed and dispatches needed actions represented by the flag into the action store.